% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyK.R
\name{graphLap}
\alias{graphLap}
\title{Plot Density by Laplace kernel.}
\usage{
graphLap(y, k, h)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}
}
\description{
Plot Kernel density by using Laplace Kernel.
}
\examples{
y<-rexp(23,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
graphLap(y,80,h)
}
