% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyK.R
\name{graphrig}
\alias{graphrig}
\title{Plot Density by RIG kernel.}
\usage{
graphrig(y, k, h)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}
}
\description{
Plot Kernel density by using Resiprocal Inverse Gaussian Kernel.
}
\examples{
y<-rexp(23,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
graphrig(y,80,h)
}
