% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyK.R
\name{mserig}
\alias{mserig}
\title{Calculate Mean Square Error( MSE) when RIG kernel is used.}
\usage{
mserig(y, k, h, type)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}

\item{type}{mention distribution of vector.If exponential distribution then use "Exp".
if use gamma distribution then use "Gamma".}
}
\value{
MSE
}
\description{
Calculate MSE by using Resiprocal Inverse Gaussian Kernel.
}
\examples{
y<-rexp(100,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
mserig(y,200,h,"exp")
}
