% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyK.R
\name{NSR}
\alias{NSR}
\title{Calculate Bandwidth proposed by Silverman for non-normal data.}
\usage{
NSR(y)
}
\arguments{
\item{y}{a numeric vector of positive values.}
}
\value{
h
}
\description{
Calculate Bandwidth proposed by Silverman for non-normal data.
}
\examples{
y <- rexp(10, 1)
 NSR(y)
}
\references{
Silverman, B. W. 1986. \emph{Density Estimation}. Chapman & Hall/ CRC, London.
}
\author{
Javaria Ahmad Khan, Atif Akbar.
}
