% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyK.R
\name{righcomp}
\alias{righcomp}
\title{Calculate MSE with and ranking of Bandwidth with respect to MSE for RIG kernel.}
\usage{
righcomp(y, k, type)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{type}{mention distribution of vector.If exponential distribution then use "Exp".
if use gamma distribution then use "Gamma".If Weibull distribution then use "Weibull".}
}
\value{
MSE with 19 bandwidths, Ranks, Minimum MSE, Maximum MSE
}
\description{
Caculate MSE with 19 bandwidths by using Resiprocal Inverse Gaussian Kernel.
}
\details{
This function helps to calculate MSE by using 19 different bandwidths which are Normal Scale Rule (NSR), Complete Cross Validation (CCV), Biased Cross Validation (BCV), Unbiased Cross Validation (UBCV),
Direct Plug-In (DPI), Modified Cross Validation (MCV), Maximum Likelihood Cross Validation (MLCV), Trimmed Cross Validation (TCV),Smooth Cross Validation (SCV), Bootstrap without Sampling (bWOs), Bootstrap with Sampling (bWs),
 Bandwidth of Altman and Leger (AL),One-sided Cross Validation (OCV), Akaike information criterion (AIC),Indirect Cross Validation (ICV), Mallow’ Cp (MallowCp), Generalized Cross Validation (GCV), Polansky and Baker Plug-In (PB),
 and Gasser, Kniep, and Köhler Cross Validation (GKK). For Laplace kernel see \code{\link{laphcomp}}
}
\examples{
 \donttest{y<-rexp(100,1)
  righcomp(y, 200, "Exp")}
}
\references{
Scaillet, O. 2004. Density estimation using inverse and reciprocal inverse Gaussian kernels. 	 \emph{Nonparametric Statistics}, \strong{16}, 217-226.
}
\author{
Javaria Ahmad Khan, Atif Akbar.
}
