% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyK.R
\name{laphcomp}
\alias{laphcomp}
\title{Calculate MSE with and ranking of Bandwidth with respect to MSE for Laplace Kernel.}
\usage{
laphcomp(y, k, type)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{type}{mention distribution of vector.If exponential distribution then use "Exp".
if use gamma distribution then use "Gamma".If Weibull distribution then use "Weibull".}
}
\value{
MSE with 19 bandwidths, Ranks, Minimum MSE, Maximum MSE
}
\description{
Caculate MSE with 19 bandwidths by using Laplace Kernel.
}
\details{
This function helps to calculate MSE by using 19 different bandwidths which are Normal Scale Rule (NSR), Complete Cross Validation (CCV), Biased Cross Validation (BCV), Unbiased Cross Validation (UBCV),
Direct Plug-In (DPI), Modified Cross Validation (MCV), Maximum Likelihood Cross Validation (MLCV), Trimmed Cross Validation (TCV), Smooth Cross Validation (SCV), Bootstrap without Sampling (bWOs), Bootstrap with Sampling (bWs),
 Bandwidth of Altman and Leger (AL), One-sided Cross Validation (OCV), Akaike information criterion (AIC), Indirect Cross Validation (ICV), Mallow’ Cp (MallowCp), Generalized Cross Validation (GCV), Polansky and Baker Plug-In (PB),
 and Gasser, Kniep, and Köhler Cross Validation (GKK). For RIG kernel see \code{\link{righcomp}}.
}
\examples{
 \donttest{y<-rexp(100,1)
  laphcomp(y, 200, "Exp")}
}
\references{
Khan, J. A.; Akbar, A. Density Estimation by Laplace Kernel. \emph{Working paper,  Department of Statistics, Bahauddin Zakariya University, Multan, Pakistan.}
}
\author{
Javaria Ahmad Khan, Atif Akbar.
}
