% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyK.R
\name{MSE}
\alias{MSE}
\title{Calculate Mean Squared Error( MSE) by using different Kernels}
\usage{
MSE(kernel, type)
}
\arguments{
\item{kernel}{type of kernel which is to be used}

\item{type}{mention distribution of vector.If exponential distribution then use \code{"Exp"}.
If use gamma distribution then use \code{"Gamma"}.If Weibull distribution then use \code{"Weibull"}.}
}
\value{
Mean Squared Error (MSE)
}
\description{
This function calculates the mean squared error (MSE) by using user specified kernel.This function is same as provided in package "DELTD". For details see \url{https://CRAN.R-project.org/package=DELTD}.
}
\references{
\url{https://CRAN.R-project.org/package=DELTD}
}
\author{
Javaria Ahmad Khan, Atif Akbar.
}
