% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsyK.R
\name{righcomp}
\alias{righcomp}
\title{Calculate MSE with and ranking of Bandwidth with respect to MSE for RIG kernel.}
\usage{
righcomp(y, k, type)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{type}{mention distribution of vector.If exponential distribution then use "Exp".
if use gamma distribution then use "Gamma".If Weibull distribution then use "Weibull".}
}
\value{
MSE withh 19 bandwidths, Ranks, Minimum MSE, Maximum MSE
}
\description{
Caculate MSE with 19 bandwidths by using Resiprocal Inverse Gaussian Kernel.
}
\examples{
 \donttest{y<-rexp(100,1)
  righcomp(y, 200, "Exp")}
}
